+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to the "note-organizer" extension will be documented in this
 
 ## [Unreleased]
 
+## [0.2.0]
+
+### Added
+ - New feature which let you import all of the .txt/.md file after project creation
+
+### Changed
+ - Update create/import project icon
+
 ## [0.1.0]
 
 Initial release